Fixing docs, regarding python2 compatibility (#79)
Qgrid is still compatible with Python 2. I mistakenly removed that from the docs because the sample notebook wasn't working with python 2, since it was saved with a python 3 kernel. To prevent other people from running into that problem I've added a separate version of the sample notebook that I saved with a python 2 kernel, and updated the docs accordingly.