JupyterLab 2 compatebility
- Update the python package coupling to the JupyterLab extension to be
less strict, allowing for a coupling to `^1.1.2` instead of exactly
`^1.1.1`. This is very relevant if we make a small bump to the extension
but don't want to make a new release of qgrid, or if qgrid would be
installed by conda and it's availability was delayed so the newer npm
package was available but the python package was outdated and coupled to
the old versions specifically.
- Move various loaders that are only used by webpack to build the json
files from being dependencies to being devDependencies.
- Bump versions of various dependencies to avoid security issues found
in packages by npm.